scene_graph_prediction/requirements.txt
Cython==0.29.23
tensorboard==2.4.1
setuptools==59.5.0
protobuf==3.17.1
open3d==0.9.0.0
scipy==1.6.3
scikit-learn==0.24.2
plyfile==0.7.4
trimesh==2.35.39
matplotlib==3.4.2
opencv-python==4.5.2.52
tqdm==4.61.0
msgpack-numpy==0.4.7.1
lmdb==1.2.1
h5py==3.3.0
json-tricks==3.15.5
pytorch-lightning==1.6.0
transformers==4.12.5
timm==0.4.12
rich==12.0.1
pyvis==0.1.9